Reorganize FE Homepage components
