+++ frontend/src/components/Homepage/Banner/Banner.test.tsx
