+++ frontend/src/components/Homepage/Banner/Banner.tsx
