+++ frontend/src/components/Homepage/Spotlight/Spotlight.module.css
