+++ frontend/src/components/Homepage/Spotlight/Spotlight.test.tsx
