+++ frontend/src/pages/index.tsx
@@ -1,3 +1,3 @@
-import Homepage from '../components/Homepage/Homepage.lazy';
+import Homepage from '../components/Homepage/Homepage/Homepage.lazy';
 
 export default Homepage;
